modify auto lua bind

+++ Code/Server/AutoLuaBind/cpp/LuaBind_GameLogic_EffectAttrs.cpp
@@ -3,8 +3,8 @@
 #include "LogicModules/GameLogic/Scene/Effects/EffectAttrs/EffectAttrs.h"	
 #include "LogicModules/GameLogic/Scene/Effects/EffectAttrs/EffectAttrsConfig.h"	
 #include "LogicModules/GameLogic/Scene/SceneModule/SceneEffects/SceneEffects.h"	
-#include "LogicModules/GameLogic/Scene/Effects/EffectBase.h"	
-#include "LogicModules/GameLogic/Scene/Effects/EffectConfigBase.h"
+#include "LogicModules/GameLogic/Scene/Effects/EffectConfigBase.h"	
+#include "LogicModules/GameLogic/Scene/Effects/EffectBase.h"
 
 namespace SolLuaBind
 {
@@ -29,7 +29,7 @@ namespace SolLuaBind
 				{
 					sol::usertype<GameLogic::EffectAttrs> meta_table(
 						sol::constructors<				
-						GameLogic::EffectAttrs(const GameLogic::EffectConfigBase *, GameLogic::SceneEffects *, unsigned long long)
+						GameLogic::EffectAttrs(const GameLogic::EffectConfigBase *, GameLogic::SceneEffects *, uint64_t)
 						>(),
 						"__StructName__", sol::property([]() {return "EffectAttrs"; })				
 						, sol::base_classes, sol::bases<
